Merging the latest to the master.

+++ include/kinx.h
@@ -307,7 +307,7 @@ extern void ir_code_dump_one(int addr, kx_code_t *code);
 extern void ir_dump(kx_context_t *ctx);
 extern void ir_dump_fixed_code(kvec_pt(kx_code_t) *fixcode);
 extern void ir_fix_code(kx_context_t *ctx, int start);
-extern int ir_exec(kx_context_t *ctx);
+extern int64_t ir_exec(kx_context_t *ctx);
 extern void native_dump(unsigned char *bytes, int size);
 
 /* for import library */
